Changes
Made changes to the flow of the API
