+++ app.js
@@ -4,15 +4,17 @@ const bodyParser = require("body-parser");
 const morgan = require("morgan");
 const path = require("path");
 const mongoose = require("mongoose");
-const swaggerJSDoc = require("swagger-jsdoc");
+
 
 app.use(morgan("dev"));
 
 //Routes used in the API
 const userRoutes = require("./API/v1/routes/user");
+const swaggerRoutes = require("./API/v1/routes/swagger");
 
 // Creates a public route for access to documents in the "public" folder
 app.use(express.static(path.join(__dirname, "public")));
+
 
 //tell mongoose to use es6 implementation of promises
 mongoose.Promise = global.Promise;
@@ -49,41 +51,9 @@ app.use((req,res,next)=>{
     next();
 });
 
-// General definition of a Swagger document
-const swaggerDefinition = {
-    info: {
-        title: "CEII portal API",
-        version: "1.0.0",
-        description: "General description of all the routes inside the API",
-    },
-    host: "db.ceii.com.sv:52138",
-    basePath: "/API/v1",
-    schemes: ["https" , "http"],
-    securityDefinitions:{
-        Authorization: {
-            type: "apiKey",
-            name: "Authorization",
-            in: "header"
-        }
-    }
-};
-// Options for Swagger
-const options = {
-    // Imports the swaggerDefinition ^
-    swaggerDefinition: swaggerDefinition,
-    // Which places will it read
-    apis: ["./API/v1/controllers/**/*.js","./API/v1/models/*.js", "./API/v1/routes/*.js"],// pass all in array
-};
-// Initialize the Swagger document
-const swaggerSpec = swaggerJSDoc(options);
-
-// This route returns a json that the index.html of Swagger readsr
-app.get("/v1/swagger.json", (req, res, next)=> {
-    res.setHeader("Content-Type", "application/json");
-    res.send(swaggerSpec);
-});
 
 // Active endpoints for the API
 app.use("/API/v1/users", userRoutes);
+app.use("/API/v1/swagger", swaggerRoutes);
 
 module.exports = app;
